+++ src/index.js
@@ -11,6 +11,7 @@ import "./styles/forms.css";
 import "./styles/menu.css";
 import "./styles/header.css";
 import "./icomoon/style.css";
+import "./components/CheckBoxes/checkbox.css";
 
 /**
  * layout components and navigational page elements
@@ -103,3 +104,8 @@ export { FormLabel } from "./components/FormLabel/FormLabel";
  * Form RadioButton
  */
 export { FormRadioButton } from "./components/FormRadioButton/FormRadioButton";
+
+/*
+ * Check box
+ */
+export { CheckBoxes } from "./components/CheckBoxes/CheckBoxes";